modificado modelo de datos

+++ app/COMPONENTS/fastapi/app/models/models.py
@@ -4,6 +4,6 @@ class SensorData(BaseModel):
     temperature: float
     humidity: float
     light_level: int
-    pressure: int
-    ppm: float
+    pressure: float
+    ppm: int
